LayoutTests/fast/hidpi/image-srcset-intrinsic-size-expected.txt

PASS document.getElementById("foo").clientWidth==200 is true
PASS document.getElementById("foo2").clientWidth==400 is true
PASS document.getElementById("foo3").clientWidth==100 is true
PASS document.getElementById("foo4").clientWidth==200 is true
This test passes if the images are all displayed with appropriate dimensions.
 
 
 
